Move prohibition of multiple reservations to licence plate model
