+++ src/api/serializers/users/reservation_serializer.py
@@ -15,7 +15,6 @@ from src.api.serializers import (
     LicencePlateSerializer,
 )
 from src.core.serializers import APIForeignKeySerializer
-from src.users.models import User
 
 
 class GetReservationSerializer(serializers.ModelSerializer):
@@ -53,15 +52,15 @@ class PostReservationSerializer(APIForeignKeySerializer):
         from_date: datetime = data["from_date"]
         to_date: datetime = data["to_date"]
         user_id: int = data["user_id"]
-        user: User = User.objects.get(pk=user_id)
-        if not user.can_reserve(from_date, to_date):
+        lp: LicencePlate = LicencePlate.objects.get(pk=user_id)
+        if not lp.can_reserve(from_date, to_date):
             raise serializers.ValidationError(
-                "You already have a reservation that time."
+                "This licence plate already has a reservation that time."
             )
         lp = LicencePlate.objects.get(pk=data["licence_plate_id"])
         if not lp.enabled:
             raise serializers.ValidationError("Licence plate is not confirmed.")
-        if data["from_date"] > data["to_date"]:
+        if from_date > to_date:
             raise serializers.ValidationError("`fromDate` must occur before `toDate`")
         if not parking_lot_is_available(
             ParkingLot.objects.get(id=data["parking_lot_id"]),
